move test file and re-enable simugraphtests
