+++ SofaKernel/modules/SofaSimulationGraph/CMakeLists.txt
@@ -39,3 +39,7 @@ sofa_create_package_with_targets(
 # Tests
 # If SOFA_BUILD_TESTS exists and is OFF, then these tests will be auto-disabled
 cmake_dependent_option(SOFASIMULATIONGRAPH_BUILD_TESTS "Compile the automatic tests" ON "SOFA_BUILD_TESTS OR NOT DEFINED SOFA_BUILD_TESTS" OFF)
+if(SOFASIMULATIONGRAPH_BUILD_TESTS)
+    enable_testing()
+    add_subdirectory(${PROJECT_NAME}_test)
+endif()
