+++ SofaKernel/modules/SofaSimulationGraph/SofaSimulationGraph_test/DAG_test.cpp
@@ -19,7 +19,7 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <sofa/simulation/testing/Node_test.h>
+#include "Node_test.h"
 
 #include <sofa/helper/testing/BaseTest.h>
 using sofa::helper::testing::BaseTest ;
