+++ SofaKernel/modules/SofaSimulationGraph/SofaSimulationGraph_test/Node_test.cpp
@@ -25,7 +25,7 @@ using sofa::helper::testing::BaseSimulationTest ;
 #include <SofaSimulationGraph/SimpleApi.h>
 using namespace sofa::simpleapi ;
 
-#include <sofa/simulation/testing/Node_test.h>
+#include "Node_test.h"
 
 namespace sofa {
